add errors mod
